Implement connectionPoolTimeout in config, and handle connection pool timeout errors separately
Previously, large thread pools driving connections through small connection pools
ended in a high rate of ConnectionPoolTimeout errors. The only fix was to set
request timeout artificially high, which is inelegant and leads to unresponsive
applications in the event of an actual network problem.

This change not only separates connection pool timeout from request timeout,
but also adds handling (via KojijiErrorInfo) for distinguishing unexpected
exceptions in Koji communications from things like timeouts in the connection
pool. Connection pool timeouts are not really possible to eliminate in all cases...
not without making an application that will hang if the network drops. So, we
should focus on detecting errors like this that are considered temporary, and
creating applications that adjust (retrying requests, etc.)

This change should start that process...though ConnectionTimeoutException is
definitely not the only case we may want to annotate as temporary.