Add better support for images (#184)
This adds two new fields to the extra map, "osbs_build" and
"submitter", and two new fields to the extra type info map, "image" and
"icm".

Although the "image" field existed in the extra map, it did not exist
in the type info map. This also adds missing image fields, including
"index" and "parent_image_builds".

This fixes several issues with "remote-sources". The type should have
been a list and not a single object. The JSON creator also did not
work.

This adds a test for parsing the extra info from the JSON map. It
removes the installation of the Koji JSON extra map serializers from
the Koji JSON module as the implementation is incomplete and it should
not be necessary.